split results into different files
